Move pt_rBR to custom implementations to add strings
