+++ src/main/kotlin/crimera/patches/twitter/misc/bringbacktwitter/strings/StringsMap.kt
@@ -6,7 +6,6 @@ object StringsMap {
         "values-en-rGB" to en_rGB.values,
         "values-ru" to ru.values,
         "values-hi" to hi.values,
-        "values-pt-rBR" to pt_rBR.values,
         "values-tr" to tr.values,
         "values-zh-rCN" to zh_rCN.values,
         "values-zh-rTW" to zh_rTW.values,
